% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_scripts.R
\name{run_simulation}
\alias{run_simulation}
\alias{vcf_to_counts}
\title{\code{call_scripts} Call the supporting (non-R) scipts in TrackSig}
\usage{
run_simulation(simName, dataDir)

vcf_to_counts(vcf_file, cna_file = NULL, purity_file = NULL)
}
\description{
\code{vcf_to_counts} From a supplied VCF file run the necessary scripts to get corrected_vaf and make_counts output. Return a multi-slotted object with useful dataframes.
}
