% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf_simulations.R
\name{get_alexMerged}
\alias{get_alexMerged}
\alias{compute_uncorrected_ccf}
\alias{save_exposures_per_time_point}
\alias{generate_ccf_simulation}
\title{Generate simulated data for \code{TrackSig}}
\usage{
get_alexMerged(signature_file = TrackSig.options()$signature_file)

compute_uncorrected_ccf(n_alt_alleles, n_ref_alleles)

save_exposures_per_time_point(data_all_clusters, sig_names, file_path,
  bin_size = 100)

generate_ccf_simulation(n_clusters, cluster_ccfs, n_mut_per_cluster,
  sig_activities, signature_def, cluster_cna_info = list(),
  mean_depth = 100, to_file = TRUE, simulation_name = "simulation",
  outdir = ".", bin_size)
}
\arguments{
\item{n_clusters}{integer number of clusters of mutations to simulate}

\item{cluster_ccfs}{list of cluster mean values}

\item{n_mut_per_cluster}{list of number of mutations to generate per cluster}

\item{sig_activities}{list of signatures and their activities for each cluster}

\item{cluster_cna_info}{list of named list with elements "fractions", \cr
"mut_cn", "total_cn". There should be one inner list per cluster.}

\item{mean_depth}{average read depth to simulate> Defaults to 100}

\item{to_file}{logical whether to save generated simulations to a file. Defaluts to TRUE}

\item{simulation_name}{simulation name to appear in outputted file name. Defaults to "simulation"}
}
\description{
Functions required for generating simulated data for TrackSig
}
\details{
\code{get_alexMerged} <man content>
}
