% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_counts.R
\name{compute_signatures_for_all_examples}
\alias{compute_signatures_for_all_examples}
\title{\code{compute_signatures_for_all_examples} Compute mutational signatures and \cr
find changepoints}
\usage{
compute_signatures_for_all_examples(countsDir, bootstrapDir,
  samples_to_run = c())
}
\description{
\code{compute_signatures_for_all_examples} Compute mutational signatures and \cr
find changepoints
}
