% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf_simulations.R
\name{create_simulation_set_knobs}
\alias{create_simulation_set_knobs}
\title{\code{create_simulation_set_knobs} generates simulated data files for knobs simualtions (n x n)}
\usage{
create_simulation_set_knobs(outdir = "simulations", mut_per_sim = 5000,
  sim_activity_file = "annotation/sim_active_in_sample.txt",
  sim_purity_file = "annotation/sim_purity.txt",
  sim_tumortype_file = "annotation/sim_tumortypes.txt",
  signature_file = "annotation/sigProfiler_SBS_signatures.txt",
  rewrite_annotations = T, n_simulations = 25)
}
\arguments{
\item{mut_per_sim}{int. Number of mutations per simulation. Default: 5000}
}
\description{
\code{create_simulation_set_knobs} generates simulated data files for knobs simualtions (n x n)
}
