% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackSig.R
\name{detectActiveSignatures}
\alias{detectActiveSignatures}
\title{get binCounts from vcfToCounts}
\usage{
detectActiveSignatures(vcfFile, cnaFile = NULL, purity = 1,
  referenceSignatures = alex_merged, threshold = 0.05, prior = NULL,
  context = generateContext(c("CG", "TA")),
  refGenome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19)
}
\arguments{
\item{referenceSignatures}{TODO}

\item{threshold}{TODO}

\item{prior}{TODO}

\item{binCounts}{TODO}
}
\description{
get binCounts from vcfToCounts
}
