% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeTrackSig.R
\name{GenomeTrackSig}
\alias{GenomeTrackSig}
\title{Determine genomic trajectory}
\usage{
GenomeTrackSig(
  counts,
  chr_level = FALSE,
  bootstrapMethod = "None",
  bootstrapSamples = 0,
  parallelize = FALSE,
  activeInSample,
  binSize,
  referenceSignatures = GenomeTrackSig:::cosmicV3,
  refGenome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19
)
}
\arguments{
\item{counts}{data frame containing mutation counts.}

\item{chr_level}{logical whether GenomeTrackSig should be run separately on
each chromosome; default is FALSE}

\item{bootstrapMethod}{string indicating desired bootstrapping method; default 'None'.
'Mutation' samples with replacement from mutations in each bin. 'Shuffle' samples a random
order of chromosomes (not applicable when chr_level = TRUE).}

\item{bootstrapSamples}{integer denoting number of bootstrap samples to take;
default is 0}

\item{parallelize}{logical whether functions should be run in parallel;
default is FALSE. Cannot be TRUE if chr_level is FALSE and bootstrapSamples is
0.}

\item{activeInSample}{list of active signatures to fit
(must match signatures present in referenceSignatures)}

\item{binSize}{number of mutations per bin}

\item{referenceSignatures}{dataframe containing definitions of mutational
signatures.}

\item{refGenome}{BSgenome to use as reference}
}
\value{
trajectory object containing one trajectory per bootstrap sample.
}
\description{
\code{GenomeTrackSig} will take an input data frame of mutation counts, and
determine a profile over the genome for the sample, based on changepoints
found using the PELT segmentation algorithm.
}
