% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTrajectories.R
\name{getChangepointsPELT}
\alias{getChangepointsPELT}
\title{#' @name getChangepointsPELT}
\usage{
getChangepointsPELT(vcaf, countsPerBin, referenceSignatures, scoreMethod,
  binSize = 100, desiredMinSegLen = NULL)
}
\description{
#' @name getChangepointsPELT
}
