% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\docType{data}
\name{generateContext}
\alias{generateContext}
\alias{helper_functions}
\alias{list}
\alias{get_values_from_list}
\alias{toVerticalMatrix}
\alias{toHorizontalMatrix}
\alias{truncate_to_range}
\title{\code{generateContext} Generate a trinucleotide context from an alphabet. Note: this involves finding all three-member
permutations of the alphabet, which can be inconveinent for large alphabets. Nucleotides are assumed to be provided as complementary pairs,
where the first of each pair is used as the reference to build the context.}
\format{An object of class \code{result} of length 1.}
\usage{
generateContext(alphabet)

list

get_values_from_list(list, name_of_value, FUN = NULL, default = NULL,
  concat_function = "cbind")

toVerticalMatrix(L)

toHorizontalMatrix(L)

truncate_to_range(mixtures, range_)
}
\arguments{
\item{alphabet}{list of pairs of characters to create combinations of as a mutation context type}
}
\value{
data.frame containing all the possible trinucleotide contextes for a mutation in the supplied alphabet
}
\description{
Non-exported functions called by \code{TrackSig} functions. \cr
Not intended for end-user use.
}
\examples{
context <- TrackSig:::generateContext(c("CG", "TA"))
dim(context)
head(context)

}
\keyword{datasets}
