% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_all_samples.R
\name{load_all_samples}
\alias{load_all_samples}
\alias{load_samples}
\alias{load_sample}
\title{\code{load_all_samples} Load previously computed mutation count data for all samples}
\usage{
load_samples(countsDir = TrackSig.options()$DIR_COUNTS,
  bootstrapDir = TrackSig.options()$BOOTSTRAP_COUNTS, outDir = ".",
  slidingWindow = FALSE)

load_sample(sampleName, countsDir)
}
\description{
\code{load_all_samples} Load previously computed mutation count data for all samples

\code{load_sample} Load previously computed mutation count data for one sample
}
