% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_counts.R
\name{load_counts}
\alias{load_counts}
\title{\code{load_counts} Load data from mut_types, phi, quadratic phi files for one sample. \cr
Combines the functionality of extract_data_for_example and load_sample}
\usage{
load_counts(phiFile, quadphiFile, example, dir_counts, tumortypes,
  dir_results = TrackSig.options()$DIR_RESULTS, dir_create = T)
}
\description{
\code{load_counts} Load data from mut_types, phi, quadratic phi files for one sample. \cr
Combines the functionality of extract_data_for_example and load_sample
}
