% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_counts.R
\name{load_sample}
\alias{load_sample}
\title{\code{load_sample} Load previously computed mutation count data for one sample}
\usage{
load_sample(sample, countsDir, bootstrapDir, tumortypes)
}
\description{
\code{load_sample} Load previously computed mutation count data for one sample
}
