% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf_simulations.R
\name{load_sim_signatures}
\alias{load_sim_signatures}
\title{Generate simulated data for \code{TrackSig}}
\usage{
load_sim_signatures(signature_file)
}
\description{
Functions required for generating simulated data for TrackSig
}
