% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSeq.R
\name{makeSeq}
\alias{makeSeq}
\title{\code{makeSeq} output a random mRNA sequence.}
\usage{
makeSeq(len, useInit = TRUE, useStop = TRUE, alphabet = c("A", "C", "G",
  "T"), p = rep(0.25, 4), seed = Sys.time())
}
\arguments{
\item{len}{integer number of codons in the output sequence.}

\item{useInit}{(logical)  make the first codon an initiation codon. Default is \code{TRUE}.}

\item{useStop}{(logical)  make the last codon a stop codon. Default is \code{TRUE}.}

\item{alphabet}{(character vector)  the elements that are randomly concatenated. Default is A/C/G/T.}

\item{p}{(numeric vector)  Probabilites for each character to occur. Default is equiprobable.}

\item{seed}{(numeric)  the seed for the RNG. Default is \code{Sys.time()}.}
}
\value{
A string.
}
\description{
\code{makeSeq} outputs a string that contains a random sequence of characters (A/C/G/T by default). By default the first three letters are "ATG" and the last three letters are a stop-codon ("TAA", "TAG" or "TGA"). The number of characters in the output is three times the \code{len} argument, i.e. it condains \code{len} codons
}
\examples{
makeSeq(7)
makeSeq(7, p = c(0.2, 0.4, 0.4, 0.2), seed = 112358)
gsub("T", "U", makeSeq(7)) # for RNA
}
