% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NOSUCH.maf}
\alias{NOSUCH.maf}
\title{NO-SUCH data.}
\format{Example A MAF formatted file with 15 rows and 32 columns:
\describe{
  \item{exampleHUGO_Symbol}{example this is a HGNC gene symbol}
  \item{exampleENTREZ_GENE_ID}{this is an ID}
  ...
}}
\source{
\url{https://example.com/docs/dummy_formats/}
}
\description{
Description of the data ....
}
\examples{
\dontrun{
system.file("data", "NOSUCH.maf", package="rpt")
fPath <- system.file("data", "NOSUCH.maf", package="rpt")
myExample <- read.delim(fPath, stringsAsFactors = FALSE)
}
}
