% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options-TrackSig.R
\name{options-TrackSig}
\alias{options-TrackSig}
\alias{TrackSig.options}
\alias{TrackSig.getOption}
\alias{TrackSig.resetOptions}
\alias{TrackSig.printOptions}
\title{\code{options-TrackSig} TrackSig Package specific options}
\usage{
TrackSig.options(...)

TrackSig.getOption(x, default = NULL)

TrackSig.resetOptions(..., ALL = FALSE)

TrackSig.printOptions()
}
\arguments{
\item{...}{option specifications. For TrackSig.options this can be named\cr
arguments or a single named list (see \link{options})\cr
For TrackSig.resetOptions, this must be the names of the options to reset. \cr
Note that pkgmaker version >= 0.9.1 is required for this to work correctly,\cr
when options other than the default have been set after the package is loaded.}

\item{x}{a character string holding an option name.}

\item{default}{if the specified option is not set in the options list,
    this value is returned.  This facilitates retrieving an option and
    checking whether it is set and setting it separately if not.}

\item{ALL}{logical that indicates if options that are not part of the default set
of options should be removed.
Note that in \pkg{pkgmaker <= 0.9} this argument is only taken into account when
no other argument is present. This is fixed in version 0.9.1.}
}
\description{
\code{options-TrackSig} TrackSig Package specific options

\code{TrackSig.options} sets/get single or multiple options, that are \cr
specific to the TrackSig package. It behaves in the same way as \link{options}.

\code{TrackSig.getOption} returns the value of a single option, that is \cr
specific to the TrackSig package. It behaves in the same way as \cr
\link{getOption}.

\code{TrackSig.resetOptions} reset all TrackSig specific options to their \cr
default values.

\code{TrackSig.printOptions} prints all TrackSig specific options along with \cr
their default values, in a relatively compact way.
}
\section{Available options}{

\describe{

\item{compute_bootstrap}{Whether to perform bootstrapping for uncertainty calculation}

} % end description
}

