% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenomeProfile.R
\name{plotGenomeProfile}
\alias{plotGenomeProfile}
\title{Plot the genomic signature activity profile of a tumor.}
\usage{
plotGenomeProfile(trajectory, show = TRUE, chr_level, cutoff)
}
\arguments{
\item{trajectory}{a list containing named elements "mixtures",
"changepoints", and 'binData'. See @seealso \link{GenomeTrackSig}.}

\item{show}{logical whether to print the plot.}

\item{chr_level}{logical whether GenomeTrackSig was run on each chromosome
separately}

\item{cutoff}{minimum proportion of bootstrap samples that must agree on a
changepoint location in order for that changepoint to be plotted}
}
\value{
ggplot object
}
\description{
\code{plotGenomeProfile} For each bin in a set of signature
mixtures, the mixture is plotted across the genome. Provided changepoints
will be highlighted.
}
