% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpaceTrack.R
\name{plotSpaceTrajectory}
\alias{plotSpaceTrajectory}
\title{Plot the genomic trajectory of a tumor.}
\usage{
plotSpaceTrajectory(trajectory, show = TRUE, chr_level, cutoff = 0)
}
\arguments{
\item{trajectory}{a list containing named elements "mixtures",
"changepoints", and 'binData'. See @seealso \link{TrackSig}.}

\item{show}{logical whether to print the plot.}

\item{chr_level}{logical whether TrackSig was run on each chromosome
separately; default is F.}

\item{cutoff}{minimum proportion of bootstrap samples that must agree on a
changepoint location in order for that changepoint to be plotted; default is
0.}
}
\value{
ggplot object
}
\description{
\code{plotSpaceTrajectory} For each bin in a set of signature
mixtures, the mixture is plotted across the genome. Provided changepoints
will be highlighted.
}
