% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binningPCAWG.R
\name{poolSamples}
\alias{poolSamples}
\title{Summarize mutation counts across all samples of interest}
\usage{
poolSamples(archivePath, typesPath, cancerType)
}
\arguments{
\item{archivePath}{file path to folder containing all cancer samples}

\item{typesPath}{file path to dataframe (csv) containing file IDs and
corresponding cancer type}

\item{cancerType}{string denoting cancer type for which all
samples will be pooled}
}
\value{
dataframe of pooled mutation counts, csv file to working directory.
}
\description{
\code{poolSamples} groups together the mutation counts for multiple cancer
samples of the same type and returns a data frame in a format compatible with TrackSig().
}
