% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binningPCAWG.R
\name{readFormat}
\alias{readFormat}
\title{Format file of mutation counts}
\usage{
readFormat(path)
}
\arguments{
\item{path}{file path to csv of mutation counts}
}
\value{
dataframe of mutation counts
}
\description{
\code{readFormat} takes in a csv of mutation counts and returns a data frame in a
format compatible with TrackSig().
}
