% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf_simulations.R
\name{simulation_functions}
\alias{simulation_functions}
\alias{load_sim_signatures}
\alias{generate_ccf_simulation}
\alias{save_as_vcf}
\title{Generate simulated data for \code{TrackSig}}
\usage{
load_sim_signatures()

generate_ccf_simulation(n_clusters, cluster_ccfs, n_mut_per_cluster,
  sig_activities, cluster_cna_info = list(), mean_depth = 100,
  to_file = TRUE, simulation_name = "simulation")

save_as_vcf(data, filename)
}
\arguments{
\item{n_clusters}{integer number of clusters of mutations to simulate}

\item{cluster_ccfs}{list of cluster mean values}

\item{n_mut_per_cluster}{list of number of mutations to generate per cluster}

\item{sig_activities}{list of signatures and their activities for each cluster}

\item{cluster_cna_info}{list of named list with elements "fractions", \cr
"mut_cn", "total_cn". There should be one inner list per cluster.}

\item{mean_depth}{average read depth to simulate> Defaults to 100}

\item{to_file}{logical whether to save generated simulations to a file. Defaluts to TRUE}

\item{simulation_name}{simulation name to appear in outputted file name. Defaults to "simulation"}
}
\description{
Functions required for generating simulated data for TrackSig

\code{load_sim_signatures} load and collapse (merge) the following mutational\cr
signatures for simulation: SBS7 (collapsed from SBS7a, SBS7b, SBS7c, SBS7d),\cr
SBS17 (collapsed from SBS17a, SBS17b), SBS2+13(collapsed from SBS2, SBS13), \cr
SBS10 (collapsed from SBS10a, SBS10b)

\code{generate_ccf_simulation} generates simulations for mutations given\cr
clusters, their CCFs, CNAs and signatures.

\code{save_as_vcf} save data generated by generate_ccf_simulation() to a file
}
