% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf_simulations.R
\name{toHorizontalMatrix}
\alias{toHorizontalMatrix}
\title{Generate simulated data for \code{TrackSig}}
\usage{
toHorizontalMatrix(L)
}
\description{
Functions required for generating simulated data for TrackSig
}
