% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_scripts.R
\name{run_simulation}
\alias{run_simulation}
\title{\code{call_scripts} Call the supporting (non-R) scipts in TrackSig}
\usage{
run_simulation(simName, dataDir)
}
\description{
\code{call_scripts} Call the supporting (non-R) scipts in TrackSig
}
