% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mutational_signatures.R,
%   R/load_all_samples.R
\name{compute_mutational_signatures}
\alias{compute_mutational_signatures}
\alias{save_data_for_samples}
\alias{compute_signatures_for_all_examples}
\alias{compute_errorbars_for_all_examples}
\title{Functions to compute mutational signatures}
\usage{
save_data_for_samples(dir_counts = TrackSig.options()$DIR_COUNTS,
  bootstrap_counts = TrackSig.options()$BOOTSTRAP_COUNTS)

compute_signatures_for_all_examples(countsDir, bootstrapDir,
  samples_to_run = c())


  compute_errorbars_for_all_examples(bootstrap_counts = TrackSig.options()$BOOTSTRAP_COUNTS)

compute_signatures_for_all_examples(countsDir, bootstrapDir,
  samples_to_run = c())
}
\arguments{
\item{dir_counts}{Directory where previously computed mutation counts are \cr
located. Defaults to TrackSig.options()$DIR_COUNTS}

\item{bootstrap_counts}{Directory where previously computed mutation bootstrapped\cr
counts are located. Defaults toTrackSig.options()$BOOTSTRAP_COUNTS}
}
\description{
Functions to compute mutational signatures

\code{save_data_for_samples} Load previously computed mutation count data and save as RData

\code{compute_signatures_for_all_examples} Compute mutational signatures and \cr
find changepoints

\code{compute_errorbars_for_all_examples} Compute mutational signatures and \cr
find changepoints, bootstrapped to estimate certainty

\code{compute_signatures_for_all_examples} Compute mutational signatures and \cr
find changepoints
}
\note{
save_data_for_samples() DEPRICATED: use load_data_for_samples() instead
}
\seealso{
\link{options-TrackSig}
}
