% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf_simulations.R
\name{create_simulation_set}
\alias{create_simulation_set}
\title{\code{create_simulation_set} generates simulated data files}
\usage{
create_simulation_set(with_CNA = FALSE)
}
\arguments{
\item{with_CNA}{logical whether to simulate data with copy number abberations.}
}
\description{
\code{create_simulation_set} generates simulated data files
}
