% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackSig.R
\name{detectActiveSignatures}
\alias{detectActiveSignatures}
\title{get binCounts from vcfToCounts}
\usage{
detectActiveSignatures(binCounts, referenceSignatures = alex_merged,
  threshold = 0.05, prior = NULL)
}
\arguments{
\item{binCounts}{TODO}

\item{referenceSignatures}{TODO}

\item{threshold}{TODO}

\item{prior}{TODO}
}
\description{
get binCounts from vcfToCounts
}
