% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackSig.R
\name{detectActiveSignatures}
\alias{detectActiveSignatures}
\title{Get a list of signatures active in a sample}
\usage{
detectActiveSignatures(df, threshold = 0.05, prior = NULL, binSize,
  referenceSignatures = TrackSig::alex_merged,
  refGenome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19)
}
\arguments{
\item{df}{dataframe of mutation counts in 1 Mb segments across the genome}

\item{threshold}{minimum activity level that signature must have to be
detected}

\item{prior}{prior on the likelihood of observing a given signature (must
match signatures present in referenceSignatures)}

\item{binSize}{number of mutations per bin}

\item{referenceSignatures}{dataframe containing definitions of mutational
signatures.}

\item{refGenome}{BSgenome to use as reference}
}
\value{
Names of signatures active in sample.
}
\description{
\code{detectActiveSignatures} determines what signatures are active above an
activity threshold, from a list of signature definintions. To do this, a
multinomial mixture model of signature activities is fit to the collection of
all mutations in the sample.
}
