% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{helper_functions}
\alias{helper_functions}
\alias{gg_color_hue}
\alias{get_values_from_list}
\alias{toVerticalMatrix}
\alias{reset_parallize}
\alias{get_sliding_window_data}
\alias{merge_data_chunks}
\alias{get_noise_sig}
\alias{get_signatures_for_current_sample}
\alias{save_data}
\alias{toHorizontalMatrix}
\alias{gather_statistics}
\alias{extract_data_for_example}
\alias{extract_data_for_simulation}
\alias{extract_bootstrap_data_for_example}
\alias{gather_summaries_per_tissue}
\alias{add_noise}
\alias{convert_window_to_indices}
\alias{find_indices_with_intersection}
\alias{read_mixtures}
\alias{get_examples_group}
\alias{compute_mean_sd_err}
\alias{get_bootstrap_mixtures}
\alias{add_list_by_element}
\alias{not_na}
\alias{truncate_to_range}
\alias{load_annotation}
\alias{get_sample_purity}
\alias{merge_signatures}
\title{Helper functions for \code{TrackSig}}
\usage{
gg_color_hue(n)

get_values_from_list(list, name_of_value, FUN = NULL, default = NULL,
  concat_function = "cbind")

toVerticalMatrix(L)

reset_parallize()

get_sliding_window_data(data, shift, gap = 1)

merge_data_chunks(vcfData)

get_noise_sig(noise_sig)

get_signatures_for_current_sample(id, active_signatures.our_samples, alex,
  noise_sig)

save_data()

toHorizontalMatrix(L)

gather_statistics(mixtures, changepoints, tumor_id, dir_name, tumor_type,
  mixtures.rescaled = NULL)

extract_data_for_example(example, dir_counts, tumortypes,
  dir_results = DIR_RESULTS, dir_create = T)

extract_data_for_simulation(example, dir_counts, dir_results = DIR_RESULTS,
  dir_create = T)

extract_bootstrap_data_for_example(example, bootstrap_counts)

gather_summaries_per_tissue(omit_signature_information = F)

add_noise(vcfData, noise_rate = 0.05)

convert_window_to_indices(total_length, window_size, gap)

find_indices_with_intersection(covered_by_changepoints,
  intervals_for_change_histogram)

read_mixtures(m_file)

get_examples_group(list_,
  EXAMPLES_PER_GROUP = TrackSig.options()$EXAMPLES_PER_GROUP,
  group = TrackSig.options()$group)

compute_mean_sd_err(mixtures_bootstrap, sig_names, dir_name = NULL,
  descr = "", exclude_zero = F)

get_bootstrap_mixtures(bootstrap_vcfs, bootstrap_phis, alex.t, dir_name, descr,
  verbose = TRUE)

add_list_by_element(list_total_sum, list_to_add)

not_na(x)

truncate_to_range(mixtures, range_)

load_annotation(tumortype_file, signature_file, active_signatures_file)

get_sample_purity(tumor_id)

merge_signatures(mixtures, sigs_to_merge)
}
\description{
Non-exported functions called by \code{TrackSig} functions. \cr
Not intended for end-user use.

\code{gg_color_hue} Select colour hues for plotting

\code{get_values_from_list} Retrieve values by name from list

\code{toVerticalMatrix} Convert a list or vector-like object to a vertial matrix

\code{reset_parallelize} Select colour hues for plotting

\code{get_sliding_window_data} <man content>

\code{get_sliding_window_data} <man content>

\code{get_sliding_window_data} <man content>

\code{get_sliding_window_data} <man content>

\code{save_data} <man content>

\code{toHorizontalMatrix} <man content>

\code{gather_statistics} <man content>

\code{extract_data_for_example} <man content>

\code{extract_data_for_simulation} <man content>

\code{extract_bootstrap_data_for_example} <man content>

\code{gather_summaries_per_tissue} <man content>

\code{add_noise} <man content>

\code{convert_window_to_indices} <man content>

\code{find_indices_with_intersection} <man content>

\code{read_mixtures} <man content>

\code{get_examples_group} <man content>

\code{compute_mean_sd_err} <man content>

\code{get_bootstrap_mixtures} <man content>

\code{add_list_by_element} <man content>

\code{not_na} <man content>

\code{truncate_to_range} <man content>

\code{load_annotation} <man content>

\code{get_sample_purity} <man content>

\code{merge_signatures} <man content>
}
