% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_counts.R
\name{loadAndScoreIt_pcawg}
\alias{loadAndScoreIt_pcawg}
\title{\code{loadAndScoreIt_pcawg} Take an input vcf file and annotation and generate the counts data.
Create all plotting output that compute_signatures_for_all_examples does.}
\usage{
loadAndScoreIt_pcawg(vcfFile, tumortypes, acronym, cnaFile = NULL,
  purityFile = NULL, saveItermediate = F)
}
\arguments{
\item{vcfFile}{path to variant calling format (vcf) file}

\item{cnaFile}{path to copy number abberation (cna) file}

\item{purityFile}{path to sample purity file}

\item{saveIntermediate}{boolean whether to save intermediate results (mutation types)}
}
\description{
\code{loadAndScoreIt_pcawg} Take an input vcf file and annotation and generate the counts data.
Create all plotting output that compute_signatures_for_all_examples does.
}
