% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackSig.R
\name{loadAndScoreIt_pcawg}
\alias{loadAndScoreIt_pcawg}
\title{\code{loadAndScoreIt_pcawg} Take an input vcf file and annotation and generate the counts data.
Create all plotting output that compute_signatures_for_all_examples does.}
\usage{
loadAndScoreIt_pcawg(vcfFile, cnaFile = NULL, purity = NULL,
  activeInSample = c("SBS1", "SBS5"), sampleID = NULL,
  refrenceSignatures = alex,
  refGenome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19)
}
\arguments{
\item{vcfFile}{path to variant calling format (vcf) file}

\item{cnaFile}{path to copy number abberation (cna) file}

\item{purityFile}{path to sample purity file}

\item{saveIntermediate}{boolean whether to save intermediate results (mutation types)


activeInSample is list used to subset refrenceSignatures}
}
\description{
\code{loadAndScoreIt_pcawg} Take an input vcf file and annotation and generate the counts data.
Create all plotting output that compute_signatures_for_all_examples does.
}
