% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapTrack.R, R/plotting.R
\name{plotTrajectory}
\alias{plotTrajectory}
\alias{plotBootstrap}
\title{Plot the evolutionary trajectory of a tumour}
\usage{
plotBootstrap(trajectory, show = TRUE)

plotTrajectory(trajectory, linearX = T, show = T)
}
\arguments{
\item{trajectory}{a list containing named elements "mixtures", "changepoints",
and "binData". See @seealso \link{TrackSig}.}

\item{show}{logical whether to print the plot}

\item{linearX}{logical whether to plot with a linearly spaced x-axis grid, or
with binned phi values}

\item{trajectory}{a list containing named elements "mixtures", "changepoints",
and "binData". See @seealso \link{TrackSig}.}

\item{linearX}{logical whether to plot with a linearly spaced x-axis grid, or
with binned phi values}

\item{show}{logical whether to print the plot}
}
\value{
ggplot object

ggplot object
}
\description{
For each bin in a set of signature mixtures, the mixture is plotted ac ross
pseudo-time. Provided changepoints will be highlighted.

For each bin in a set of signature mixtures, the mixture is plotted ac ross
pseudo-time. Provided changepoints will be highlighted.
}
