% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf_simulations.R
\name{write_sim_annotation}
\alias{write_sim_annotation}
\title{\code{write_sim_annotation} save active signatures generated by generate_ccf_simulation() to a file
Uses PCAWG_sigProfiler signature set}
\usage{
write_sim_annotation(simulation_name, sig_activities, sig_header,
  sim_activity_file, sim_purity_file, sim_tumortype_file)
}
\description{
\code{write_sim_annotation} save active signatures generated by generate_ccf_simulation() to a file
Uses PCAWG_sigProfiler signature set
}
