% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackSig.R
\name{TrackSig}
\alias{TrackSig}
\title{\code{TrackSig} Take an input vcf file and annotation and generate the counts data.
Create all plotting output that compute_signatures_for_all_examples does.}
\usage{
TrackSig(vcfFile, activeInSample, cnaFile = NULL, purity = NULL,
  sampleID = NULL, referenceSignatures = alex_merged,
  scoreMethod = "SigFreq", binSize = 100, desiredMinSegLen = NULL,
  refGenome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19)
}
\arguments{
\item{vcfFile}{path to variant calling format (vcf) file}

\item{activeInSample}{list of signatures that are active. All listed signatures
must be present in the referenceSIgnatures dataframe.}

\item{cnaFile}{path to copy number abberation (cna) file}

\item{purity}{TODO}

\item{sampleID}{name to call sample. If none provided, name will be automatically drawn from the provided vcf file name.}

\item{referenceSignatures}{TODO}

\item{scoreMethod}{TODO}

\item{binSize}{TODO}

\item{desiredMinSegLen}{TODO}

\item{refGenome}{TODO}
}
\description{
\code{TrackSig} Take an input vcf file and annotation and generate the counts data.
Create all plotting output that compute_signatures_for_all_examples does.
}
