% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackSig.R
\name{TrackSig}
\alias{TrackSig}
\title{Determine an evolutionary trajectory.}
\usage{
TrackSig(
  df,
  binSize,
  activeInSample,
  sampleID = NULL,
  referenceSignatures = TrackSig::alex_merged,
  scoreMethod = "Signature",
  nCutoff = 10000,
  desiredMinSegLen = 1,
  refGenome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19
)
}
\arguments{
\item{df}{dataframe of mutation counts from which we can fit activity trajectories}

\item{binSize}{number of mutations per bin}

\item{activeInSample}{list of signatures names to fit the exposures of. All listed signatures must be present in the referenceSignatures dataframe.}

\item{sampleID}{name to call sample. If none provided, name will be automatically drawn from the provided vcf file name.}

\item{referenceSignatures}{dataframe containing definitions of mutational signatures.}

\item{scoreMethod}{string to indicate what scoring method to apply when finding changepoints. Default = "Signature"}

\item{nCutoff}{maximum number of total mutations to consider (samples with more than nCutoff muations will be down-sampled)}

\item{desiredMinSegLen}{minimum number of mutations to include in a PELT segment (the desiredMinSegLen will be overridden if there are too few for accurate scoring)}

\item{refGenome}{BSgenome to use as reference}
}
\description{
\code{TrackSig} will take an input VCF file and corresponding annotation, and determine an eveolutionary trajectory for the sample, based on changepoints found using the PELT segmentation algorithm.
}
